perf: Make traverseAst not copy the AST every time.
Our linters do analysis, no transformation. This adds a MapAst that does
what the old TraverseAst did, and makes TraverseAst a pure traversal
without copying.
