+++ src/Language/Cimple/IO.hs
@@ -18,12 +18,12 @@ import qualified Data.Text                       as Text
 import qualified Data.Text.Encoding              as Text
 import           Language.Cimple.Ast             (Node)
 import           Language.Cimple.Lexer           (Lexeme, runAlex)
+import           Language.Cimple.MapAst          (TextActions, mapAst,
+                                                  textActions)
 import qualified Language.Cimple.Parser          as Parser
 import           Language.Cimple.Program         (Program)
 import qualified Language.Cimple.Program         as Program
 import           Language.Cimple.TranslationUnit (TranslationUnit)
-import           Language.Cimple.TraverseAst     (TextActions, textActions,
-                                                  traverseAst)
 import qualified Language.Cimple.TreeParser      as TreeParser
 
 type StringNode = Node (Lexeme String)
@@ -31,7 +31,7 @@ type TextNode = Node (Lexeme Text)
 
 toTextAst :: [StringNode] -> [TextNode]
 toTextAst stringAst =
-    evalState (traverseAst cacheActions stringAst) Map.empty
+    evalState (mapAst cacheActions stringAst) Map.empty
   where
     cacheActions :: TextActions (State (Map String Text)) String Text
     cacheActions = textActions $ \s -> do
